+++ spec/e2e/migration_spec.rb
@@ -13,16 +13,16 @@ describe 'migration tasks' do
       class User
         include Neo4j::ActiveNode
         property :name
-        has_many :out, :songs, model_class: Song, type: 'songs'
+        has_many :out, :songs, model_class: :Song, type: 'songs'
       end
 
       class Song
         include Neo4j::ActiveNode
         property :name
 
-        has_many :in, :owners, model_class: User, origin: :songs
-        has_many :out, :singers, model_class: User, rel_class: MigrationSpecs::SecondRelClass
-        has_many :out, :new_singers, model_class: User, rel_class: MigrationSpecs::ThirdRelClass
+        has_many :in, :owners, model_class: :User, origin: :songs
+        has_many :out, :singers, model_class: :User, rel_class: MigrationSpecs::SecondRelClass
+        has_many :out, :new_singers, model_class: :User, rel_class: MigrationSpecs::ThirdRelClass
         def custom_id
           'my new id'
         end